Makes TelemetrySpi public
